new changing login

+++ database/migrations/2019_09_17_153729_create_social_facebook_accounts_table.php
@@ -14,7 +14,7 @@ class CreateSocialFacebookAccountsTable extends Migration
     public function up()
     {
         Schema::create('social_facebook_accounts', function (Blueprint $table) {
-            $table->integer('user_id');
+            $table->bigIncrements('user_id');
             $table->string('provider_user_id');
             $table->string('provider');
             $table->timestamps();
